🚚 refactor(file): ShareButton 컴포넌트 위치 이동 및 임포트 정리
- `ShareButton` 컴포넌트를 `ui/top/bar/component`에서 `ui` 패키지로 이동하여 폴더 구조 정리
- `FileScreen`에서 변경된 경로에 맞춰 `ShareButton` 임포트 호출부 수정
- `FileScreen` 내에서 사용되지 않는 `_ShareBottomSheet` 임포트 제거
